backend/requirements.txt
# =============================================
# Scientific Navigator — Direct Dependencies
# =============================================
# This file lists ONLY the packages directly imported
# by the project. Transitive deps are resolved by pip.
# =============================================

# --- Django Stack ---
Django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.0
python-dotenv>=1.0

# --- Database ---
psycopg2-binary>=2.9

# --- LangChain / RAG ---
langchain>=1.0
langchain-chroma>=1.0
langchain-community>=0.4
langchain-ollama>=1.0
langchain-text-splitters>=1.0

# --- Vector Store ---
chromadb>=1.0

# --- PDF Processing ---
pypdf>=4.0

# --- External API Connectors ---
arxiv>=2.0
biopython>=1.80
requests>=2.30

# --- Ollama Client ---
ollama>=0.5

# --- Hybrid Retrieval ---
rank-bm25>=0.2

